migration to new journal viewer and editor
Completely migrated from the old journal editor to the new one. Also added a new journal entry viewer with better design as a BottomSheet. Recordings are currently still missing from the viewer but are planned to be implemented in the next commits. A few bugs were fixed as well

+++ app/src/main/java/com/bitflaker/lucidsourcekit/main/dreamjournal/JournalInMemoryManager.java
@@ -51,11 +51,13 @@ public class JournalInMemoryManager {
         if(loadedJournalEntryIds.contains(id)) {
             int index = loadedJournalEntryIds.indexOf(id);
             JournalInMemory jim = loadedJournalEntries.get(index);
-            List<RecordingData> recData = jim.getAudioRecordings();
-            for (RecordingData rec : recData) {
-                File file = new File(rec.getFilepath());
-                if(file.exists()) {
-                    file.delete();
+            if(jim.getEditMode() == JournalInMemory.EditMode.CREATE){
+                List<RecordingData> recData = jim.getAudioRecordings();
+                for (RecordingData rec : recData) {
+                    File file = new File(rec.getFilepath());
+                    if(file.exists()) {
+                        file.delete();
+                    }
                 }
             }
             loadedJournalEntryIds.remove(index);
